pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "writer-reader"
version = "0.1.0"
description = "Segment-based book manuscript editor with git integration"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Bob Matsuoka", email = "bob@matsuoka.com" }
]
keywords = ["writing", "manuscript", "editor", "markdown", "git"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
]
requires-python = ">=3.10"
dependencies = [
    "flask>=2.0",
    "flask-cors>=3.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
writer-reader = "writer_reader.cli:main"

[project.urls]
Homepage = "https://github.com/bobmatnyc/writer-reader"
Repository = "https://github.com/bobmatnyc/writer-reader"
Issues = "https://github.com/bobmatnyc/writer-reader/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/writer_reader"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["writer_reader"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/writer_reader"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
